chore: configure theme setup
