+++ src/__stories__/CheckboxInput.tsx
@@ -6,9 +6,11 @@ import { withKnobs } from '@storybook/addon-knobs';
 import { boolean, text } from '@storybook/addon-knobs/react';
 
 import { CheckboxInput } from '../components/CheckboxInput';
+import { ThemeZone } from '../theme';
 
 storiesOf('Inputs', module)
   .addDecorator(withKnobs)
+  .addDecorator(storyFn => <ThemeZone name="formtron">{storyFn()}</ThemeZone>)
   .add('CheckboxInput', () => {
     return (
       <CheckboxInput
