+++ src/__stories__/IntegerInput.tsx
@@ -6,9 +6,11 @@ import { withKnobs } from '@storybook/addon-knobs';
 import { boolean, number, text } from '@storybook/addon-knobs/react';
 
 import { IntegerInput } from '../components/IntegerInput';
+import { ThemeZone } from '../theme';
 
 storiesOf('Inputs', module)
   .addDecorator(withKnobs)
+  .addDecorator(storyFn => <ThemeZone name="formtron">{storyFn()}</ThemeZone>)
   .add('IntegerInput', () => {
     return (
       <IntegerInput
