--- src/components/theme.ts
@@ -1,16 +1,0 @@
-import { createThemedModule, ICustomTheme } from '@stoplight/ui-kit';
-
-export type themeZones = 'formtron' | string;
-
-export type themeTypes = IFormtronTheme;
-
-export interface IFormtronTheme extends ICustomTheme {
-  canvas?: {
-    bg: string;
-    fg: string;
-    valid: string;
-    invalid: string;
-  };
-}
-
-export const { useTheme, ThemeZone } = createThemedModule<themeZones, themeTypes>();
