+++ src/theme.ts
@@ -1,6 +1,16 @@
-import { createThemedModule } from '@stoplight/ui-kit';
-import { IFormtronTheme } from './types';
+import { createThemedModule, ICustomTheme } from '@stoplight/ui-kit';
 
 export type themeZones = 'formtron' | string;
 
-export const { useTheme, ThemeZone } = createThemedModule<themeZones, IFormtronTheme>();
+export type themeTypes = IFormtronTheme;
+
+export interface IFormtronTheme extends ICustomTheme {
+  canvas?: {
+    bg: string;
+    fg: string;
+    valid: string;
+    invalid: string;
+  };
+}
+
+export const { useTheme, ThemeZone } = createThemedModule<themeZones, themeTypes>();
